% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_pred.R
\name{average_pred}
\alias{average_pred}
\title{Ensemble of regression models}
\usage{
average_pred(X_train, y_train, X_test)
}
\value{
A vector of predictions
}
\description{
This gives an mathematical mean of different model predictions
}
