% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_pred_without_KNN.R
\name{average_pred_without_KNN}
\alias{average_pred_without_KNN}
\title{Ensemble of regression models (the same as average_pred() but without KNN)}
\usage{
average_pred_without_KNN(X_train, y_train, X_test)
}
\arguments{
\item{X_train}{training predictor vector}

\item{y_train}{training response vector}

\item{X_test}{predictor vector}
}
\value{
A vector of predictions
}
\description{
When given training predictors, training responses, and predictors, this function
outputs a vector of predictions
The models used for this ensemble are: Random forest, LASSO, Ridge and linear
}
