% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{combineDistn.fn}
\alias{combineDistn.fn}
\title{A simple way to sample from multiple models with no correlation between parameters}
\usage{
combineDistn.fn(totSamps, means, stds, wts = 1)
}
\arguments{
\item{totSamps}{Total number of samples across all models}

\item{means}{The means of the parameter for each model}

\item{stds}{The standard deviations of the parameter for each model}

\item{wts}{The weight for each model.  A single value is equal weighting. Standardized to 1.}
}
\description{
A simple way to sample from multiple models with no correlation between parameters
}

