% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{cor2cov.fn}
\alias{cor2cov.fn}
\title{convert a correlation matrix to a covariance matrix}
\usage{
cor2cov.fn(corMat, sds)
}
\arguments{
\item{corMat}{The correlation matrix}

\item{sds}{A vector of the standard deviations associated with the diagonal of the correlation matrix}
}
\description{
convert a correlation matrix to a covariance matrix
}

