% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSample.R
\name{createSample.fn}
\alias{createSample.fn}
\title{Create a multivariate normal sample
Checks for positive definite covariance matrix and makes a correction if not}
\usage{
createSample.fn(n, means, Sigma)
}
\arguments{
\item{n}{Sample size}

\item{means}{A vector of means}

\item{Sigma}{Variance-Covariance matrix}
}
\description{
Create a multivariate normal sample
Checks for positive definite covariance matrix and makes a correction if not
}

