% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enbagg.R
\name{enbagg}
\alias{enbagg}
\title{Bagging Ensemble Learning Function}
\usage{
enbagg(data, split, learn, resp_var, tree)
}
\arguments{
\item{dataset, }{Split ratio, Number of Learners, Response variable, Number of Trees}
}
\value{
Average Accuracy
}
\description{
This function takes 1) dataset, 2) Split ratio,3) Number of samples on which models to be train, 4) Response variable as arguments, 5) Numbers of trees to train
It functions takes out samples from the dataset provided as per the Learner number passed in the argument
Each sample will be train using Random Forest with number of trees passed in argument
This function returns Average Accuracy
}
