\name{ens.rpart}
\alias{ens.rpart}
\title{modified version of rpart}
\description{
  slightly modified version of \code{\link{rpart}} used by \code{\link{ensemble}}
}
\usage{
ens.rpart(formula, data=NULL, weights, na.action=na.rpart, method, model=FALSE,
x=FALSE, y=TRUE, parms, control, ...)
}
\seealso{\code{\link{rpart}}}
\keyword{rpart}
