\name{ensemble-package}
\alias{ensemble-package}
\alias{ensemble}
\docType{package}
\title{
Fit ensemble models and optimize tuning parameters
}
\description{
This package provides functions for building models via cross-validation and combining those models via ensembling to generate a good prediction on test data.  Additionally, the package provides a function for optimizing tuning parameters.
}
\details{
\tabular{ll}{
Package: \tab ensemble\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-18\cr
}
}
\author{
Maintainer: Josh Browning <rockclimber112358@gmail.com>
}
\keyword{ package }
\keyword{ ensemble }
\keyword{ cross-validation }
\keyword{ tuning-parameter }
\examples{
~~ simple examples of the most important functions ~~
}
