% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_plot.R
\name{error_plot}
\alias{error_plot}
\title{Graph performance of ensemble of regression models}
\usage{
error_plot(X_train, y_train, X_test, y_test, rfntree, KNNk, rfmtry, doKNN)
}
\arguments{
\item{X_train}{training predictor vector}

\item{y_train}{training response vector}

\item{X_test}{predictor vector}

\item{y_test}{responses vector}

\item{rfntree}{number of trees to grow. This should not be set to too small a number. Set to 100 by default}

\item{KNNk}{Number of neighbours considered it KNN, set to 5 if unspecified}

\item{rfmtry}{Number of variables randomly sampled as candidates at each split in random forest. Set to 6 by default}

\item{doKNN}{Whether to run KNN or not, TRUE by default}
}
\value{
A vector of predictions
}
\description{
When given training predictors, training responses, and predictors, responses this function
outputs a plot mean squire errors of the models
The models used for this ensemble are: KNN, Random forest, LASSO, Ridge and linear
}
