% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssEA.R
\name{ESEA_fast}
\alias{ESEA_fast}
\title{Enhancer Set Enrichment Analysis (ESEA) using fgsea}
\usage{
ESEA_fast(weight, compare.list, minSize = 100, nPermSimple = 10000)
}
\arguments{
\item{weight}{A named numeric vector of weights.}

\item{compare.list}{A list of enhancer sets (or pathways) to compare. Each element should be a vector of enhancer identifiers.}

\item{minSize}{Integer. Minimum number of enhancers required for a set to be considered (default is 100).}

\item{nPermSimple}{Integer. The number of permutations to perform in the FGSEA algorithm (default is 10000).}
}
\value{
A data frame with FGSEA results that includes columns for the normalized enrichment score (NES), p-value,
  q-value, a signed q-value (calculated as the sign of NES times -log10(q-value)), enrichment score (ES),
  the number of more extreme permutations, the size of the enhancer set, and the leading edge.
}
\description{
Performs enhancer set enrichment analysis (ESEA) on differential enhancer expression (DEE) results using the fgsea package.
}
\details{
The function first creates a named weight vector using the sign of the DEE statistic multiplied by the negative log10 of the p-value.
FGSEA is then performed using the \code{fgseaMultilevel} function from the fgsea package. The results are reformatted,
with columns renamed, and a signed q-value is computed. Finally, the results are ordered by the signed q-value in descending order.
}
