% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{filter_geneHancer}
\alias{filter_geneHancer}
\title{Filter Intragenic GeneHancers}
\usage{
filter_geneHancer(
  GeneHancer.file,
  GeneCode.file,
  extend.bp = 0,
  filter.promoter = TRUE,
  filter.exon = TRUE
)
}
\arguments{
\item{GeneHancer.file}{A character string specifying the path to the GeneHancer annotation file.}

\item{GeneCode.file}{A character string specifying the path to the GTF file (gene annotations).}

\item{extend.bp}{An integer specifying the number of base pairs to extend exon regions (default is 0).}

\item{filter.promoter}{Logical. If TRUE, filters out GeneHancer entries classified as promoters (default is TRUE).}

\item{filter.exon}{Logical. If TRUE, filters out GeneHancer entries that overlap with extended exon regions (default is TRUE).}
}
\value{
A data frame containing the filtered GeneHancer annotations.
}
\description{
Filters out intragenic GeneHancers by excluding those overlapping with extended exon regions,
and optionally filters out promoter elements. Both \code{GeneHancer.file} and \code{GeneCode.file}
must be based on the same genome build.
}
