% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCov.R
\name{getCov.fn}
\alias{getCov.fn}
\title{Get the covariance matrix from a SS model for a specific set of parameters
Uses the SS version of the correlation matrix}
\usage{
getCov.fn(ssModel, vars)
}
\arguments{
\item{ssModel}{A SS model loaded using SS_output from r4ss}

\item{vars}{Names of the variables for the covariance matrix}
}
\description{
Get the covariance matrix from a SS model for a specific set of parameters
Uses the SS version of the correlation matrix
}

