% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVals.R
\name{getVals.fn}
\alias{getVals.fn}
\title{Gets the mean and standard deviation of parameters from SS output}
\usage{
getVals.fn(x, param, cols = c("Value", "StdDev"))
}
\arguments{
\item{x}{The matrix of parameters or derived parameters from SS_output}

\item{param}{The name of the parameter}

\item{cols}{The column names for the mean and standard deviation, respectively}
}
\description{
Gets the mean and standard deviation of parameters from SS output
}

