% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{readCor.fn}
\alias{readCor.fn}
\title{To Read the admb .cor file}
\usage{
readCor.fn(corFile, preNames = 4)
}
\arguments{
\item{corFile}{The location and name of the .cor file}

\item{preNames}{The number of columns before the correlations are reported}
}
\description{
To Read the admb .cor file
}

