% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_analysis.R
\name{viper_enhancer_score}
\alias{viper_enhancer_score}
\title{Compute VIPER Enhancer Scores}
\usage{
viper_enhancer_score(
  cpmMatrix,
  enhancersets,
  cores = 1,
  pleiotropy = TRUE,
  pleiotropyArgs = list(regulators = 0.05, shadow = 0.05, targets = 10, penalty = 20,
    method = "adaptive")
)
}
\arguments{
\item{cpmMatrix}{A numeric matrix representing counts per million (CPM) expression values,
where rows represent enhancers and columns represent samples.}

\item{enhancersets}{A list in which each element is a vector of enhancers for a specific context.}

\item{cores}{Integer. Number of cores to use for parallel computation (default is 1). Parallel computation can significantly reduce the time for pleiotropy (might need to adjust the cores if face certain errors).}

\item{pleiotropy}{Logical. Whether to apply pleiotropy correction (default is TRUE).}

\item{pleiotropyArgs}{A list of parameters for pleiotropy correction. Default is
\code{list(regulators = 0.05, shadow = 0.05, targets = 10, penalty = 20, method = "adaptive")}.}
}
\value{
A VIPER score matrix as returned by the \code{viper} function from the VIPER package.
}
\description{
Computes VIPER enhancer scores by creating a regulon object from enhancer sets
and applying the VIPER algorithm to the provided CPM matrix.
}
\details{
The function constructs a regulon object by assigning a weight of 1 and equal confidence to each
target gene in the provided enhancer sets. It then computes the VIPER score using \code{viper::viper()}.
Pleiotropy correction is highly recommended if some enhancer sets share large number of enhancers.
}
