% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_pred.R
\name{average_pred}
\alias{average_pred}
\title{Ensemble of regression models}
\usage{
average_pred(X_train, y_train, X_test, rfntree, KNNk, rfmtry, doKNN)
}
\arguments{
\item{X_train}{training predictor vector}

\item{y_train}{training response vector}

\item{X_test}{predictor vector}

\item{rfntree}{number of trees to grow. This should not be set to too small a number}
}
\value{
A vector of predictions
}
\description{
When given training predictors, training responses, and predictors, this function
outputs a vector of predictions
The models used for this ensemble are: KNN, Random forest, LASSO, Ridge and linear
}
